ETL/Python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
grpcio==1.59.3
grpcio-tools==1.59.3
paho-mqtt==1.6.1
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.7
redis==5.0.1
boto3==1.34.0
botocore==1.34.0
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
aioredis==2.0.1
asyncio-mqtt==0.16.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# AWS Tools
awscli==1.32.0
boto3==1.34.0
aws-cdk-lib==2.108.0
constructs==10.3.0

# Development Tools
docker==6.1.3
docker-compose==1.29.2
terraform==1.6.0
